Reorganizing files. Adding library PFSP_nodes
