+++ baselines/pfsp/pfsp_gpu_cuda.c
@@ -18,7 +18,7 @@
 #include "lib/c_bound_simple.h"
 #include "lib/c_bound_johnson.h"
 #include "lib/c_taillard.h"
-#include "evaluate.h"
+#include "lib/evaluate.h"
 
 /******************************************************************************
 CUDA error checking
@@ -32,24 +32,7 @@ void gpuAssert(cudaError_t code, const char *file, int line, bool abort) {
   }
 }
 
-/*******************************************************************************
-Implementation of PFSP Nodes.
-*******************************************************************************/
-
-//TO DO: create separate files that deal with nodes and another one that deals with pools
-
-// BLOCK_SIZE, MAX_JOBS and struct Node are defined in parameters.h
-
-// Initialization of nodes is done by CPU only
-
-void initRoot(Node* root, const int jobs)
-{
-  root->depth = 0;
-  root->limit1 = -1;
-  for (int i = 0; i < jobs; i++) {
-    root->prmu[i] = i;
-  }
-}
+/* //TO DO: create separate files that deal with nodes and another one that deals with pools */
 
 /*******************************************************************************
 Implementation of a dynamic-sized single pool data structure.
